admin operations v6
We have deleted the articles we added from the database. We sent the deleted recycle using Laravel's soft delete structure so that the data we deleted is not destroyed directly. We have performed any recovery or operations remaining after removing it from the recycle bin. In addition, we destroyed all the files belonging to the deleted article.